third_party/blink/web_tests/fast/events/constructors/media-encrypted-event-constructor-expected.txt

This tests the constructor for the MediaEncryptedEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new MediaEncryptedEvent('eventType').bubbles is false
PASS new MediaEncryptedEvent('eventType').cancelable is false
PASS new MediaEncryptedEvent('eventType').initDataType is ""
PASS new MediaEncryptedEvent('eventType').initData is null
PASS new MediaEncryptedEvent('eventType', { bubbles: true, cancelable: true }).bubbles is true
PASS new MediaEncryptedEvent('eventType', { bubbles: true, cancelable: true }).cancelable is true
PASS new MediaEncryptedEvent('eventType', { bubbles: true, cancelable: true }).initDataType is ""
PASS new MediaEncryptedEvent('eventType', { bubbles: true, cancelable: true }).initData is null
PASS new MediaEncryptedEvent('eventType', { initDataType: 'foo' }).initDataType is "foo"
PASS new MediaEncryptedEvent('eventType', { initData: initData }).initData is initData
PASS new MediaEncryptedEvent('eventType', { initData: 'foo' }).initData threw exception TypeError: Failed to construct 'MediaEncryptedEvent': Failed to read the 'initData' property from 'MediaEncryptedEventInit': Failed to convert value to 'ArrayBuffer'..
PASS new MediaEncryptedEvent('eventType', { initData: 1 }).initData threw exception TypeError: Failed to construct 'MediaEncryptedEvent': Failed to read the 'initData' property from 'MediaEncryptedEventInit': Failed to convert value to 'ArrayBuffer'..
PASS successfullyParsed is true

TEST COMPLETE
